requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.1
scikit-learn==1.2.2
seaborn==0.12.2
tensorboard==2.12.1
tensorflow-intel==2.12.0
